Extend logging to our multiprocess Process's
Extend the logging configuration to the multiprocessing processes so
that we can capture logs from them in Python as well. Processes create
an entirely new Python process so can't inherit their configuration from
the root logger. Fortunately, we can pass log messages back on a queue
and let the main Process continue to handle them in its own thread. This
would be nice to simplify at some point.

This change also adds a way to create a new log file for each job run.
This necessitated pulling our log level and format configuration out
into their own functions so we kept that configuration in a single
place.