Fix the test imports
I also had to add an __init__.py to keep pytest from compaining about
similarly named unit test files. In the future, it'd be good to change
to name our tests after the file they're testing in the src directory.

I've also moved the actively used data files to a "testdata" directory.
I'd like to cleanup the rest of the data files in the repo if we can
determine they're unused.