Create new processes with spawn instead of fork
Fork can be dangerous as it copies parent processes variables,
including lock states but not the threads that can unlock them. Python
is changing this default in 3.14. However, for now we have to manually
specify that we want to "spawn" new processes with a clean slate instead
of "fork"-ing them.

This was causing deadlock errors with our logging thread. There's some
debate if we should even be using a thread in the main process while
we're using multiprocessing elsewhere.