Add actionable Trivy output
The Trivy Sarif output is useful for uploading to the GitHub Security
tab but isn't friendly for developers to debug as the reason for the CI
failure is hidden from in the build output. Add a second Trivy scan to
generate the useful "table" formatted output, and relegate the Sarif
output to be behind-the-scenes.