Update the python tooling in the correct location
Updating the python tooling after activating the virtual environment
will update the venv version of the tooling but not the underlying
OS/system version of the tooling. We may need to update this for the OS
image and then for the venv to cover all our bases. However, for now I
think we can do this by updating the environment in the builder layer
which gets copied into the virtual environment, and by updating the prod
image before we activate the virtual environment.